Reorganize files + WIP TypeScript conversion
god dang it did I forget to push the changelog and package.json version bump for the last release? I am a dumb
